fix bug

+++ backend/consumer-server/src/main/java/com/grp202116/consumerserver/controller/ProjectController.java
@@ -6,6 +6,7 @@ import com.grp202116.consumerserver.pojo.ModelDO;
 import com.grp202116.consumerserver.pojo.ProjectDO;
 import com.grp202116.consumerserver.util.ExportUtils;
 import com.grp202116.consumerserver.util.FileUtils;
+import org.apache.ibatis.jdbc.Null;
 import org.springframework.core.io.ByteArrayResource;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
@@ -118,8 +119,10 @@ public class ProjectController {
         FileUtils.deleteDirectory(projectPath);
         ModelDO model = modelMapper.getByProjectId(projectId);
 
-        FileUtils.deleteFile(model.getModelPath());
-        FileUtils.deleteFile(model.getLabelsPath());
+        if(model != null){
+            FileUtils.deleteFile(model.getModelPath());
+            FileUtils.deleteFile(model.getLabelsPath());
+        }
 
         projectMapper.deleteByProjectId(projectId);
     }
